requirements.txt
    aiogram==2.14.3
    aiohttp==3.7.4.post0
    async-timeout==3.0.1
    attrs==21.2.0
    Babel==2.9.1
    certifi==2021.5.30
    chardet==4.0.0
    charset-normalizer==2.0.4
    constants==0.6.0
    html==1.13
    idna==3.2
    multidict==5.1.0
    pip==21.1.2
    python-dotenv==0.19.0
    pytz==2021.1
    requests==2.26.0
    setuptools==57.0.0
    telebot==0.0.4
    telegram==0.0.1
    typing-extensions==3.10.0.0
    urllib3==1.26.6
    wheel==0.36.2
    yarl==1.6.3
